Split nssdb storage in more manageable modules
