+++ src/storage/nssdb/mod.rs
@@ -8,19 +8,21 @@ use std::sync::{Arc, Mutex};
 use crate::attribute::{AttrType, Attribute, CkAttrs};
 use crate::error::{Error, Result};
 use crate::interface::*;
-use crate::kasn1::oid::*;
-use crate::kasn1::pkcs::*;
 use crate::misc::copy_sized_string;
 use crate::object::Object;
 use crate::storage;
 use crate::storage::sqlite_common::check_table;
 use crate::storage::{Storage, StorageDBInfo, StorageTokenInfo};
 use crate::token::TokenFacilities;
-use crate::{byte_ptr, sizeof, void_ptr};
 
 use rusqlite::types::{FromSqlError, Value, ValueRef};
 use rusqlite::{Connection, Rows};
 use zeroize::Zeroize;
+
+mod ci;
+mod config;
+use ci::*;
+use config::*;
 
 impl From<std::fmt::Error> for Error {
     fn from(_: std::fmt::Error) -> Error {
@@ -77,496 +79,6 @@ fn nss_col_to_type(col: &str) -> Result<CK_ULONG> {
 }
 
 static NSS_PASS_CHECK: &[u8; 14] = b"password-check";
-
-#[derive(
-    asn1::Asn1Read, asn1::Asn1Write, PartialEq, Hash, Clone, Eq, Debug,
-)]
-pub struct BrokenAlgorithmIdentifier<'a> {
-    pub oid: asn1::DefinedByMarker<asn1::ObjectIdentifier>,
-    #[defined_by(oid)]
-    pub params: BrokenAlgorithmParameters<'a>,
-}
-
-#[derive(
-    asn1::Asn1DefinedByRead,
-    asn1::Asn1DefinedByWrite,
-    PartialEq,
-    Eq,
-    Hash,
-    Clone,
-    Debug,
-)]
-pub enum BrokenAlgorithmParameters<'a> {
-    #[defined_by(PBES2_OID)]
-    Pbes2(BrokenPBES2Params<'a>),
-    #[defined_by(PBMAC1_OID)]
-    Pbmac1(PBMAC1Params<'a>),
-
-    #[defined_by(AES_128_CBC_OID)]
-    Aes128Cbc(&'a [u8]),
-    #[defined_by(AES_256_CBC_OID)]
-    Aes256Cbc(&'a [u8]),
-    #[defined_by(HMAC_WITH_SHA256_OID)]
-    HmacWithSha256(Option<asn1::Null>),
-}
-
-#[derive(
-    asn1::Asn1Read, asn1::Asn1Write, PartialEq, Eq, Hash, Clone, Debug,
-)]
-pub struct BrokenPBES2Params<'a> {
-    pub key_derivation_func: Box<AlgorithmIdentifier<'a>>,
-    pub encryption_scheme: Box<BrokenAlgorithmIdentifier<'a>>,
-}
-
-#[derive(
-    asn1::Asn1Read, asn1::Asn1Write, PartialEq, Eq, Hash, Clone, Debug,
-)]
-struct NSSEncryptedDataInfo<'a> {
-    pub algorithm: Box<BrokenAlgorithmIdentifier<'a>>,
-    pub enc_or_sig_data: &'a [u8],
-}
-
-fn pbkdf2_derive(
-    facilities: &TokenFacilities,
-    params: &PBKDF2Params,
-    secret: &[u8],
-    key_template: &[CK_ATTRIBUTE],
-) -> Result<Object> {
-    let mech = facilities.mechanisms.get(CKM_PKCS5_PBKD2)?;
-
-    let ck_params = CK_PKCS5_PBKD2_PARAMS2 {
-        saltSource: CKZ_SALT_SPECIFIED,
-        pSaltSourceData: void_ptr!(params.salt.as_ptr()),
-        ulSaltSourceDataLen: CK_ULONG::try_from(params.salt.len())?,
-        iterations: CK_ULONG::try_from(params.iteration_count)?,
-        prf: match params.prf.oid() {
-            &HMAC_WITH_SHA1_OID => CKP_PKCS5_PBKD2_HMAC_SHA1,
-            &HMAC_WITH_SHA256_OID => CKP_PKCS5_PBKD2_HMAC_SHA256,
-            _ => return Err(CKR_MECHANISM_PARAM_INVALID)?,
-        },
-        pPrfData: std::ptr::null_mut(),
-        ulPrfDataLen: 0,
-        pPassword: byte_ptr!(secret.as_ptr()),
-        ulPasswordLen: CK_ULONG::try_from(secret.len())?,
-    };
-
-    mech.generate_key(
-        &CK_MECHANISM {
-            mechanism: CKM_PKCS5_PBKD2,
-            pParameter: void_ptr!(&ck_params),
-            ulParameterLen: sizeof!(CK_PKCS5_PBKD2_PARAMS2),
-        },
-        key_template,
-        &facilities.mechanisms,
-        &facilities.factories,
-    )
-}
-
-fn aes_cbc_decrypt(
-    facilities: &TokenFacilities,
-    key: &Object,
-    iv: &[u8],
-    data: &[u8],
-) -> Result<Vec<u8>> {
-    let mut adj_iv_vec: Vec<u8>;
-
-    /* NSS has a Broken IV in the encoded data, so we need to adjust
-     * the IV we get from the decoding and prepend it with 0x04 0x0E
-     * which are thebytes that make a 16 bytes buffer "look like" a
-     * DER encoded OCTET_STRING. */
-    let adj_iv = match iv.len() {
-        14 => {
-            adj_iv_vec = Vec::with_capacity(16);
-            adj_iv_vec.extend_from_slice(&[4, 14]);
-            adj_iv_vec.extend_from_slice(iv);
-            adj_iv_vec.as_slice()
-        }
-        16 => iv,
-        _ => return Err(CKR_ARGUMENTS_BAD)?,
-    };
-
-    let ck_mech = CK_MECHANISM {
-        mechanism: CKM_AES_CBC_PAD,
-        pParameter: void_ptr!(adj_iv),
-        ulParameterLen: CK_ULONG::try_from(adj_iv.len())?,
-    };
-    let mech = facilities.mechanisms.get(CKM_AES_CBC_PAD)?;
-    let mut op = mech.decryption_new(&ck_mech, key)?;
-    let mut plain = vec![0u8; op.decryption_len(data.len(), false)?];
-    let len = op.decrypt(data, &mut plain)?;
-    plain.resize(len, 0);
-    Ok(plain)
-}
-
-pub fn decrypt_data(
-    facilities: &TokenFacilities,
-    secret: &[u8],
-    data: &[u8],
-) -> Result<Vec<u8>> {
-    let info = match asn1::parse_single::<NSSEncryptedDataInfo>(data) {
-        Ok(i) => i,
-        Err(e) => {
-            print!("{}", e);
-            return Err(CKR_DATA_INVALID)?;
-        }
-    };
-
-    let pbes2 = match &info.algorithm.params {
-        BrokenAlgorithmParameters::Pbes2(ref params) => params,
-        _ => return Err(CKR_MECHANISM_INVALID)?,
-    };
-
-    let key = match &pbes2.key_derivation_func.params {
-        AlgorithmParameters::Pbkdf2(ref params) => {
-            let ck_class: CK_OBJECT_CLASS = CKO_SECRET_KEY;
-            let ck_key_type: CK_KEY_TYPE = CKK_AES;
-            let ck_key_len: CK_ULONG = match params.key_length {
-                Some(l) => CK_ULONG::try_from(l)?,
-                None => return Err(CKR_MECHANISM_PARAM_INVALID)?,
-            };
-            let ck_true: CK_BBOOL = CK_TRUE;
-            let mut key_template = CkAttrs::with_capacity(5);
-            key_template.add_ulong(CKA_CLASS, &ck_class);
-            key_template.add_ulong(CKA_KEY_TYPE, &ck_key_type);
-            key_template.add_ulong(CKA_VALUE_LEN, &ck_key_len);
-            key_template.add_bool(CKA_DECRYPT, &ck_true);
-            key_template.add_bool(CKA_ENCRYPT, &ck_true);
-
-            pbkdf2_derive(facilities, params, secret, key_template.as_slice())?
-        }
-        _ => return Err(CKR_MECHANISM_INVALID)?,
-    };
-    match &pbes2.encryption_scheme.params {
-        BrokenAlgorithmParameters::Aes128Cbc(ref iv) => {
-            aes_cbc_decrypt(facilities, &key, iv, info.enc_or_sig_data)
-        }
-        BrokenAlgorithmParameters::Aes256Cbc(ref iv) => {
-            aes_cbc_decrypt(facilities, &key, iv, info.enc_or_sig_data)
-        }
-        _ => Err(CKR_MECHANISM_INVALID)?,
-    }
-}
-
-/* NSS has a hardcoded list of attributes that are authenticated,
- * some are vendor defined attributes */
-const AUTHENTICATED_ATTRIBUTES: [CK_ATTRIBUTE_TYPE; 2] = [
-    CKA_MODULUS,
-    CKA_PUBLIC_EXPONENT,
-    //CKA_CERT_SHA1_HASH,
-    //CKA_CERT_MD5_HASH,
-    //CKA_TRUST_SERVER_AUTH,
-    //CKA_TRUST_CLIENT_AUTH,
-    //CKA_TRUST_EMAIL_PROTECTION,
-    //CKA_TRUST_CODE_SIGNING,
-    //CKA_TRUST_STEP_UP_APPROVED,
-    //CKA_NSS_OVERRIDE_EXTENSIONS,
-];
-
-fn hmac_verify(
-    facilities: &TokenFacilities,
-    mechanism: CK_MECHANISM_TYPE,
-    key: &Object,
-    nssobjid: u32,
-    sdbtype: u32,
-    plaintext: &[u8],
-    signature: &[u8],
-) -> Result<()> {
-    let siglen = CK_ULONG::try_from(signature.len())?;
-    let ck_mech = CK_MECHANISM {
-        mechanism: mechanism,
-        pParameter: void_ptr!(&siglen),
-        ulParameterLen: sizeof!(CK_ULONG),
-    };
-    let mech = facilities.mechanisms.get(mechanism)?;
-    let mut op = mech.verify_new(&ck_mech, key)?;
-    let objid = nssobjid.to_be_bytes();
-    op.verify_update(&objid)?;
-    let attrtype = sdbtype.to_be_bytes();
-    op.verify_update(&attrtype)?;
-    op.verify_update(plaintext)?;
-    op.verify_final(signature)
-}
-
-fn check_signature(
-    facilities: &TokenFacilities,
-    secret: &[u8],
-    attribute: &[u8],
-    signature: &[u8],
-    nssobjid: u32,
-    sdbtype: u32,
-) -> Result<()> {
-    let info = match asn1::parse_single::<NSSEncryptedDataInfo>(signature) {
-        Ok(i) => i,
-        Err(e) => {
-            print!("{}", e);
-            return Err(CKR_DATA_INVALID)?;
-        }
-    };
-
-    let pbmac1 = match &info.algorithm.params {
-        BrokenAlgorithmParameters::Pbmac1(ref params) => params,
-        _ => return Err(CKR_MECHANISM_INVALID)?,
-    };
-
-    let key = match &pbmac1.key_derivation_func.params {
-        AlgorithmParameters::Pbkdf2(ref params) => {
-            let ck_class: CK_OBJECT_CLASS = CKO_SECRET_KEY;
-            let ck_key_type: CK_KEY_TYPE = CKK_GENERIC_SECRET;
-            let ck_key_len: CK_ULONG = match params.key_length {
-                Some(l) => CK_ULONG::try_from(l)?,
-                None => return Err(CKR_MECHANISM_PARAM_INVALID)?,
-            };
-            let ck_true: CK_BBOOL = CK_TRUE;
-            let mut key_template = CkAttrs::with_capacity(5);
-            key_template.add_ulong(CKA_CLASS, &ck_class);
-            key_template.add_ulong(CKA_KEY_TYPE, &ck_key_type);
-            key_template.add_ulong(CKA_VALUE_LEN, &ck_key_len);
-            key_template.add_bool(CKA_SIGN, &ck_true);
-            key_template.add_bool(CKA_VERIFY, &ck_true);
-
-            pbkdf2_derive(facilities, params, secret, key_template.as_slice())?
-        }
-        _ => return Err(CKR_MECHANISM_INVALID)?,
-    };
-
-    match &pbmac1.message_auth_scheme.params {
-        AlgorithmParameters::HmacWithSha256(None) => hmac_verify(
-            facilities,
-            CKM_SHA256_HMAC_GENERAL,
-            &key,
-            nssobjid,
-            sdbtype,
-            attribute,
-            info.enc_or_sig_data,
-        ),
-        _ => Err(CKR_MECHANISM_INVALID)?,
-    }
-}
-
-fn enckey_derive(
-    facilities: &TokenFacilities,
-    pin: &[u8],
-    salt: &[u8],
-) -> Result<Vec<u8>> {
-    let mech = facilities.mechanisms.get(CKM_SHA_1)?;
-    let mut op = mech.digest_new(&CK_MECHANISM {
-        mechanism: CKM_SHA_1,
-        pParameter: std::ptr::null_mut(),
-        ulParameterLen: 0,
-    })?;
-    op.digest_update(salt)?;
-    op.digest_update(pin)?;
-    let mut digest = vec![0u8; op.digest_len()?];
-    op.digest_final(digest.as_mut_slice())?;
-    Ok(digest)
-}
-
-/* Mostly documented at:
- * https://nss-crypto.org/reference/security/nss/legacy/pkcs11/module_specs/index.html
- */
-#[derive(Debug)]
-struct NSSConfig {
-    configdir: Option<String>,
-    updatedir: Option<String>,
-    updateid: Option<String>,
-    secmod: String,
-    manufacturer: Option<String>,
-    library_description: Option<String>,
-    cert_prefix: String,
-    key_prefix: String,
-    token_crypto_description: Option<String>,
-    token_db_description: Option<String>,
-    token_fips_description: Option<String>,
-    slot_crypto_description: Option<String>,
-    slot_db_description: Option<String>,
-    slot_fips_description: Option<String>,
-    update_description: Option<String>,
-    min_pin_len: usize,
-    read_only: bool,
-    no_mod_db: bool,
-    no_cert_db: bool,
-    no_key_db: bool,
-    force_open: bool,
-    password_required: bool,
-    optimize_space: bool,
-    /* tokens: currently unsupported */
-}
-
-impl Default for NSSConfig {
-    fn default() -> Self {
-        NSSConfig {
-            configdir: None,
-            updatedir: None,
-            updateid: None,
-            secmod: String::from("secmod.db"),
-            manufacturer: None,
-            library_description: None,
-            cert_prefix: String::from(""),
-            key_prefix: String::from(""),
-            token_crypto_description: None,
-            token_db_description: None,
-            token_fips_description: None,
-            slot_crypto_description: None,
-            slot_db_description: None,
-            slot_fips_description: None,
-            update_description: None,
-            min_pin_len: 0,
-            read_only: false,
-            no_mod_db: false,
-            no_cert_db: false,
-            no_key_db: false,
-            force_open: false,
-            password_required: false,
-            optimize_space: false,
-        }
-    }
-}
-
-impl NSSConfig {
-    fn parse_flags(&mut self, args: &[u8]) -> Result<()> {
-        let mut idx = 0;
-        while idx < args.len() {
-            let next = match args[idx..].iter().position(|&x| x == ',' as u8) {
-                Some(n) => n,
-                None => args.len(),
-            };
-            let flag = String::from_utf8_lossy(&args[idx..next]).to_lowercase();
-            match flag.as_str() {
-                "readonly" => self.read_only = true,
-                "nomoddb" => self.no_mod_db = true,
-                "nocertdb" => self.no_cert_db = true,
-                "nokeydb" => self.no_key_db = true,
-                "forceopen" => self.force_open = true,
-                "passwordrequired" => self.password_required = true,
-                "optimizespace" => self.optimize_space = true,
-                _ => return Err(CKR_ARGUMENTS_BAD)?,
-            }
-            idx = next + 1;
-        }
-        Ok(())
-    }
-
-    fn parse_parameter(&mut self, args: &[u8]) -> Result<usize> {
-        let name: String;
-        let value: String;
-
-        /* find param name */
-        let mut idx = match args.iter().position(|&x| x == '=' as u8) {
-            Some(x) => x,
-            None => Err(CKR_ARGUMENTS_BAD)?,
-        };
-
-        if args.len() <= idx + 2 {
-            return Err(CKR_ARGUMENTS_BAD)?;
-        }
-
-        name = String::from_utf8_lossy(&args[0..idx]).to_lowercase();
-
-        let find = match char::from(args[idx + 1]) {
-            '\'' => b'\'',
-            '\"' => b'\"',
-            '(' => b')',
-            '{' => b'}',
-            '[' => b']',
-            '<' => b'>',
-            _ => b' ',
-        };
-        let valx = if find != b' ' { idx + 2 } else { idx + 1 };
-        idx = valx;
-
-        while idx < args.len() {
-            if let Some(pos) = args[idx..].iter().position(|&x| x == find) {
-                idx = pos;
-
-                /* backtrack check for escapes */
-                let mut esc = 0;
-                while esc < pos {
-                    if args[pos - 1 - esc] == '\\' as u8 {
-                        esc += 1;
-                    } else {
-                        break;
-                    }
-                }
-                if esc % 2 == 1 {
-                    idx += 1;
-                    /* escaped */
-                    continue;
-                }
-                break;
-            } else {
-                idx = args.len();
-            }
-        }
-        if idx >= args.len() {
-            /* This may be the last parameter, in which case it is ok
-             * if not trailing space is present otherwise error out */
-            if idx == args.len() && find != ' ' as u8 {
-                return Err(CKR_ARGUMENTS_BAD)?;
-            }
-        }
-
-        value = String::from_utf8_lossy(&args[valx..idx]).to_string();
-
-        if idx < args.len() {
-            idx += 1;
-        }
-
-        match name.as_str() {
-            "configdir" => self.configdir = Some(value),
-            "updatedir" => self.updatedir = Some(value),
-            "updateid" => self.updateid = Some(value),
-            "secmod" => self.secmod = value,
-            "manufacturerid" => self.manufacturer = Some(value),
-            "librarydescription" => self.library_description = Some(value),
-            "certprefix" => self.cert_prefix = value,
-            "keyprefix" => self.key_prefix = value,
-            "cryptotokendescription" => {
-                self.token_crypto_description = Some(value)
-            }
-            "dbtokendescription" => self.token_db_description = Some(value),
-            "fipstokendescription" => self.token_fips_description = Some(value),
-            "cryptoslotdescription" => {
-                self.slot_crypto_description = Some(value)
-            }
-            "dbslotdescription" => self.slot_db_description = Some(value),
-            "fipsslotdescription" => self.slot_fips_description = Some(value),
-            "updatetokendescription" => self.update_description = Some(value),
-            "minpwlen" => self.min_pin_len = value.parse::<usize>()?,
-            "flags" => self.parse_flags(value.as_bytes())?,
-            _ => return Err(CKR_ARGUMENTS_BAD)?,
-        }
-
-        Ok(idx)
-    }
-
-    /* parse nss configuration string */
-    fn from_args(args: &str) -> Result<NSSConfig> {
-        let mut config: NSSConfig = Default::default();
-
-        let bargs = args.as_bytes();
-        let mut idx = 0usize;
-
-        while idx < bargs.len() {
-            idx = config.parse_parameter(&bargs[idx..])?;
-        }
-        Ok(config)
-    }
-
-    #[cfg(feature = "fips")]
-    fn get_token_label_as_bytes(&self) -> &[u8] {
-        match self.token_fips_description {
-            Some(ref s) => s.as_bytes(),
-            None => storage::TOKEN_LABEL.as_bytes(),
-        }
-    }
-    #[cfg(not(feature = "fips"))]
-    fn get_token_label_as_bytes(&self) -> &[u8] {
-        match self.token_db_description {
-            Some(ref s) => s.as_bytes(),
-            None => storage::TOKEN_LABEL.as_bytes(),
-        }
-    }
-}
 
 struct NSSSearchQuery {
     certs: Option<String>,
